requirements.txt
Django==4.2.1
django-crispy-forms==1.11.2
psycopg2-binary==2.8.5
dj-database-url==0.5.0
